requirements.txt
fastapi
uvicorn[standard]
pydantic
torch
sentence-transformers
faiss-cpu
networkx
pandas
rank_bm25
nltk
ragas
python-multipart
tf-keras
neo4j
